Backend/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==23.2
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
